+++ src/api/configuration.rs
@@ -9,20 +9,20 @@ macro_rules! api_url (
 
 pub(crate) use api_url;
 
-pub struct ImgurHandle {
+pub struct ImgurClient {
     pub client_id: String,
     pub client: Client,
 }
 
-impl fmt::Debug for ImgurHandle {
+impl fmt::Debug for ImgurClient {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "ImgurClient - client_id: {}", self.client_id)
     }
 }
 
-impl ImgurHandle {
+impl ImgurClient {
     pub fn new(client_id: String) -> Self {
         let client = Client::new();
-        ImgurHandle { client_id, client }
+        ImgurClient { client_id, client }
     }
 }