+++ src/api/delete_image.rs
@@ -1,4 +1,4 @@
-use crate::api::configuration::{api_url, ImgurHandle};
+use crate::api::configuration::{api_url, ImgurClient};
 
 use super::send_api_request;
 
@@ -6,7 +6,7 @@ use anyhow::Error as anyhow_err;
 use reqwest::Method;
 use std::io::{Error, ErrorKind};
 
-pub async fn delete_image(c: ImgurHandle, delete_hash: String) -> Result<String, anyhow_err> {
+pub async fn delete_image(c: ImgurClient, delete_hash: String) -> Result<String, anyhow_err> {
     let uri = api_url!(format!("image/{delete_hash}"));
     let res = send_api_request(&c, Method::DELETE, uri, None).await?;