+++ src/api/get_image.rs
@@ -1,4 +1,4 @@
-use crate::api::configuration::{api_url, ImgurHandle};
+use crate::api::configuration::{api_url, ImgurClient};
 use crate::api::ImageInfo;
 
 use super::send_api_request;
@@ -7,7 +7,7 @@ use anyhow::Error as anyhow_err;
 use reqwest::Method;
 use std::io::{Error, ErrorKind};
 
-pub async fn get_image(c: ImgurHandle, image: &str) -> Result<ImageInfo, anyhow_err> {
+pub async fn get_image(c: ImgurClient, image: &str) -> Result<ImageInfo, anyhow_err> {
     let uri = api_url!(format!("image/{image}"));
     let res = send_api_request(&c, Method::GET, uri, None).await?;