+++ src/api/mod.rs
@@ -7,6 +7,6 @@ pub mod get_image;
 pub mod rate_limit;
 pub mod upload_image;
 
-pub use configuration::ImgurHandle;
+pub use configuration::ImgurClient;
 pub use image_type::*;
 pub use send_request::send_api_request;