+++ src/cli/credits.rs
@@ -1,17 +1,11 @@
-use imgurs::api::{rate_limit::rate_limit, ImgurHandle};
+use imgurs::api::{rate_limit::rate_limit, ImgurClient};
 
 use chrono::{prelude::DateTime, Utc};
-use log::{error, info};
-use std::{
-    process::exit,
-    time::{Duration, UNIX_EPOCH},
-};
+use log::info;
+use std::time::{Duration, UNIX_EPOCH};
 
-pub async fn credits(client: ImgurHandle) {
-    let i = rate_limit(client).await.unwrap_or_else(|e| {
-        error!("{e}");
-        exit(1);
-    });
+pub async fn credits(client: ImgurClient) {
+    let i = rate_limit(client).await.expect("send api request");
 
     let d = UNIX_EPOCH + Duration::from_secs(i.data.user_reset.try_into().unwrap());
     let datetime = DateTime::<Utc>::from(d);