+++ src/cli/delete_image.rs
@@ -1,13 +1,10 @@
-use imgurs::api::{configuration::ImgurHandle, delete_image::delete_image as del_img};
+use imgurs::api::{delete_image::delete_image as del_img, ImgurClient};
 
-use log::{error, info};
-use std::process::exit;
+use log::info;
 
-pub async fn delete_image(client: ImgurHandle, delete_hash: String) {
-    let i = del_img(client, delete_hash).await.unwrap_or_else(|e| {
-        error!("{e}");
-        exit(1);
-    });
-
+pub async fn delete_image(client: ImgurClient, delete_hash: String) {
+    let i = del_img(client, delete_hash)
+        .await
+        .expect("send api request");
     info!("{i}");
 }