+++ src/cli/info_image.rs
@@ -1,15 +1,8 @@
-use imgurs::api::{configuration::ImgurHandle, get_image::get_image};
+use imgurs::api::{get_image::get_image, ImgurClient};
 
 use super::print_image_info;
 
-use log::error;
-use std::process::exit;
-
-pub async fn image_info(client: ImgurHandle, id: &str) {
-    let i = get_image(client, id).await.unwrap_or_else(|e| {
-        error!("{e}");
-        exit(1);
-    });
-
+pub async fn image_info(client: ImgurClient, id: &str) {
+    let i = get_image(client, id).await.expect("send api request");
     print_image_info(i, false);
 }