+++ src/config/toml.rs
@@ -2,14 +2,13 @@ use super::Config;
 
 use anyhow::Error;
 use dirs::config_dir;
-use log::{error, warn};
+use log::warn;
 use std::{
     fs::{create_dir_all, read_to_string, File},
     io::Write as _,
     path::Path,
-    process::exit,
 };
-use toml::from_str;
+use toml::from_str as toml_from_str;
 
 const CONFIG_DIR: &str = "/imgurs/config.toml";
 
@@ -19,12 +18,12 @@ pub fn parse() -> Config {
 
         let default_config = include_str!(concat!("../../config.toml"));
 
-        let sys_config_dir = config_dir().unwrap();
+        let sys_config_dir = config_dir().expect("find config dir");
         let config_dir = format!("{}{CONFIG_DIR}", sys_config_dir.to_string_lossy());
         let config_path = Path::new(&config_dir);
 
         let prefix = config_path.parent().unwrap();
-        create_dir_all(prefix).unwrap();
+        create_dir_all(prefix).expect("create config dir");
 
         let mut file_ref = File::create(config_path).expect("create failed");
 
@@ -32,21 +31,15 @@ pub fn parse() -> Config {
             .write_all(default_config.as_bytes())
             .expect("failed write default config to file");
 
-        toml().unwrap_or_else(|e| {
-            error!("parse toml config: {e}");
-            exit(3);
-        })
+        toml().expect("parse toml config")
     })
 }
 
 fn toml() -> Result<Config, Error> {
     let config_dir = config_dir().unwrap();
-
-    let file_dir: String = String::from(config_dir.to_string_lossy()) + CONFIG_DIR;
-
+    let file_dir = format!("{}{CONFIG_DIR}", config_dir.to_string_lossy());
     let toml_str = read_to_string(file_dir).map_err(Error::new)?;
-
-    let decode = from_str(&toml_str).map_err(Error::new)?;
+    let decode = toml_from_str(&toml_str).map_err(Error::new)?;
 
     Ok(decode)
 }